exception handled

+++ src/main/java/com/arun/Employee_crud/service/EmployeeServiceImpl.java
@@ -1,5 +1,7 @@
 package com.arun.Employee_crud.service;
 
+import com.arun.Employee_crud.exceptions.EmployeeAlreadyExists;
+import com.arun.Employee_crud.exceptions.EmployeeNotFound;
 import com.arun.Employee_crud.model.Employee;
 import com.arun.Employee_crud.repository.EmployeeRepo;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,30 +18,54 @@ public class EmployeeServiceImpl implements EmployeeService{
 
     @Override
     public List<Employee> findAll() {
+
+
         return employeeRepository.findAll();
     }
 
     @Override
-    public List<Employee> search(String q) {
-        return employeeRepository.findByNameContaining(q);
+    public Optional<List<Employee>> search(String q) {
+        if (employeeRepository.findByNameContaining(q).isEmpty()) {
+            throw new EmployeeNotFound("Employee with name " + q + " not found");
+        }
+        return Optional.ofNullable(employeeRepository.findByNameContaining(q));
     }
     @Override
     public Optional<Employee> findOne(int id) {
-          return employeeRepository.findById(id);
+          return Optional.ofNullable(employeeRepository.findById(id).orElseThrow(
+                  () -> new EmployeeNotFound("Employee with id " + id + " not found")
+          ));
 
     }
 
     @Override
-    public void save(Employee contact) {
+    public String save(Employee contact) {
+         Employee email = employeeRepository.findByEmail(contact.getEmail());
+            if (email != null) {
+                throw new EmployeeAlreadyExists("Employee with email " + contact.getEmail() + " already exists");
+            }
+            employeeRepository.save(contact);
+            return "Employee Created Successfully!";
+
+
+    }
+
+    @Override
+    public String update(Employee contact) {
         employeeRepository.save(contact);
+        return "Employee Updated Successfully!";
     }
 
     @Override
-    public void delete(int id) {
+    public String delete(int id) {
+        if (employeeRepository.findById(id).isEmpty()) {
+            throw new EmployeeNotFound("Employee with id " + id + " not found");
+        }
         employeeRepository.deleteById(id);
+        return "Employee Deleted Successfully!";
     }
     @Override
     public boolean existsByEmail(String email) {
-        return employeeRepository.findByEmail(email) != null;
+        return employeeRepository.findByEmail(email) == null;
     }
 }
